feat : implement the invoice reminder email sending

+++ src/invoice/application/use-cases/delete-invoice.usecase.ts
@@ -1,10 +1,14 @@
 import { getInjection } from "@/di/container";
+import { InvoiceNotFoundException } from "../execeptions/specific.exception";
 
 export const deleteInvoiceUseCase = {
     async execute(invoiceId: string) {
         const { authenticationService, invoiceRepository } = this.getServices();
         await authenticationService.verifySessionUser();
-
+        const invoice = invoiceRepository.get(invoiceId);
+        if (!invoice) {
+            throw new InvoiceNotFoundException();
+        }
         await invoiceRepository.remove(invoiceId);
     },
     getServices() {
